requirements.txt
pandas
numpy
streamlit
transformers
faiss-cpu
sentence-transformers                                                                                  
beautifulsoup4
python-dotenv
langchain==0.1.1
python-dotenv
openai
